api/requirements.txt
flask
flask-cors
mysql-connector-python
requests
pillow
numpy